// src/app/anomaly/model/label-type.enum.ts
export enum LabelTypeEnum {
  SUBSCRIPTION = 'subscription',
  CONSUMPTION = 'consumption',
  TURPE = 'turpe',
  DATE = 'date',
  INDEX = 'index',
  UNIT_PRICE = 'unit_price',
  AMOUNT = 'amount',
  DELIVERY_POINT_CHANGE = 'delivery_point_change',
  NONE = 'none',
  CLIENT = 'client',
  PROVIDER = 'provider'
}